DataEngineeringLearning/requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.12.2
plotly==5.18.0
dash==2.16.1
dash-bootstrap-components==1.5.0
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2
jupyterlab==4.0.9
notebook==7.0.6
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
celery==5.3.4
apache-airflow==2.8.0
great-expectations==0.18.8
mlflow==2.9.2
wandb==0.16.1
optuna==3.5.0
shap==0.44.1
lime==0.2.0.1
evidently==0.4.13
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
pre-commit==3.6.0
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
joblib==1.3.2
dask==2023.12.1
